Refactor: split sections into separate files
