+++ src/components/sections/ar/ARinput.jsx
@@ -1,9 +1,14 @@
+import React from 'react';
 import { FormControl, Grid, Input, InputAdornment, InputLabel } from '@material-ui/core';
 import { Attribute, Item } from 'gw2-ui-bulk';
-import React from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import { changeAR, changeOmnipotion, getAR, getOmniPotion } from '../../state/gearOptimizerSlice';
-import CheckboxComponent from '../baseComponents/CheckboxComponent';
+import {
+  changeAR,
+  changeOmnipotion,
+  getAR,
+  getOmniPotion,
+} from '../../../state/gearOptimizerSlice';
+import CheckboxComponent from '../../baseComponents/CheckboxComponent';
 
 const ARinput = () => {
   const dispatch = useDispatch();
