+++ src/components/sections/forcedslots/ForcedSlots.jsx
@@ -3,10 +3,10 @@ import Autocomplete from '@material-ui/lab/Autocomplete';
 import { Item } from 'gw2-ui-bulk';
 import React from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import { changeForcedSlot, getGeneric, getPriority } from '../../state/gearOptimizerSlice';
-import { GEAR_SLOTS } from '../../utils/gw2-data';
-import { firstUppercase } from '../../utils/usefulFunctions';
-import { AFFIXES } from './Affixes';
+import { changeForcedSlot, getGeneric, getPriority } from '../../../state/gearOptimizerSlice';
+import { GEAR_SLOTS } from '../../../utils/gw2-data';
+import { firstUppercase } from '../../../utils/usefulFunctions';
+import { AFFIXES } from '../Affixes';
 
 const styles = (theme) => ({
   textField: { marginTop: 0, marginBottom: 0 },
