+++ src/components/sections/skills/Skills.jsx
@@ -2,8 +2,8 @@ import { Typography, withStyles } from '@material-ui/core';
 import { Skill } from 'gw2-ui-bulk';
 import React from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import { addSkill, getSkills, removeSkill } from '../../state/gearOptimizerSlice';
-import CheckboxComponent from '../baseComponents/CheckboxComponent';
+import { addSkill, getSkills, removeSkill } from '../../../state/gearOptimizerSlice';
+import CheckboxComponent from '../../baseComponents/CheckboxComponent';
 
 const styles = (theme) => ({
   text: {
